+++ src/controllers/sector/index.ts
@@ -11,12 +11,17 @@ export const sectorHandler = {
         success: true,
         data: { sectors },
       },
-      StatusCodes.OK
+      StatusCodes.OK,
     );
   },
 
   async getById(c: Context) {
     const id = c.req.param("id");
+
+    if (!id) {
+      throw new BadRequestError("id is required");
+    }
+
     const sector = await sectorService.getSectorById(id);
 
     return c.json(
@@ -24,7 +29,7 @@ export const sectorHandler = {
         success: true,
         data: { sector },
       },
-      StatusCodes.OK
+      StatusCodes.OK,
     );
   },
 
@@ -39,13 +44,17 @@ export const sectorHandler = {
         message: "Sector created successfully",
         data: { sector },
       },
-      StatusCodes.CREATED
+      StatusCodes.CREATED,
     );
   },
 
   async update(c: Context) {
     const id = c.req.param("id");
     const { name, description } = await c.req.json();
+
+    if (!id) {
+      throw new BadRequestError("id is required");
+    }
 
     const sector = await sectorService.updateSector(id, name, description);
 
@@ -55,25 +64,39 @@ export const sectorHandler = {
         message: "Sector updated successfully",
         data: { sector },
       },
-      StatusCodes.OK
+      StatusCodes.OK,
     );
   },
 
   async delete(c: Context) {
     const id = c.req.param("id");
+
+    if (!id) {
+      throw new BadRequestError("id is required");
+    }
+
     await sectorService.deleteSector(id);
+
+    if (!id) {
+      throw new BadRequestError("id is required");
+    }
 
     return c.json(
       {
         success: true,
         message: "Sector deleted successfully",
       },
-      StatusCodes.OK
+      StatusCodes.OK,
     );
   },
 
   async getStats(c: Context) {
     const id = c.req.param("id");
+
+    if (!id) {
+      throw new BadRequestError("id is required");
+    }
+
     const stats = await sectorService.getSectorStats(id);
 
     return c.json(
@@ -81,7 +104,7 @@ export const sectorHandler = {
         success: true,
         data: { stats },
       },
-      StatusCodes.OK
+      StatusCodes.OK,
     );
   },
 };
